+++ config/default.js
@@ -338,6 +338,22 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 		},
 
 		/**
+		 * Returns true if all assets from the initial entry point have to inline within HTML files
+		 *
+		 * @cli inline-initial
+		 * @env INLINE_INITIAL
+		 *
+		 * @returns {boolean}
+		 */
+		inlineInitial() {
+			return o('inline-initial', {
+				env: true,
+				type: 'boolean',
+				default: false
+			});
+		},
+
+		/**
 		 * Some webpack options to optimize build
 		 */
 		optimize: {
@@ -488,6 +504,10 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 				}
 
 				return concatUrls(...args);
+			}
+
+			if (pathVal) {
+				return concatUrls(pathVal, '/').replace(/^[/]+/, '/');
 			}
 
 			return pathVal;