+++ src/super/i-static-page/modules/ss-helpers/helpers.js
@@ -18,7 +18,7 @@ exports.needInline = needInline;
  * @returns {boolean}
  */
 function needInline(forceInline) {
-	return Boolean(webpack.fatHTML() || forceInline);
+	return Boolean(webpack.fatHTML() || webpack.inlineInitial() || forceInline);
 }
 
 exports.addPublicPath = addPublicPath;
@@ -32,7 +32,7 @@ exports.addPublicPath = addPublicPath;
 function addPublicPath(path) {
 	const
 		staticExpr = `concatURLs(${toExpr(webpack.publicPath())}, ${toExpr(path)})`,
-		concatURLs = "function concatURLs(a, b) { return a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, ''); }";
+		concatURLs = "function concatURLs(a, b) { return a ? a.replace(/[\\\\/]+$/, '') + '/' + b.replace(/^[\\\\/]+/, '') : b; }";
 
 	if (webpack.dynamicPublicPath()) {
 		const